- Fixed some blurriness when rendering centered text using FTGL - Changed the browser to use a proper font for item names - Opening the preferences or base resource archive dialogs from the map editor window menu will no longer raise the main window when said dialog is closed - Don't draw round thing shadows when thing_drawtype is 2 (sprites)
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@1175 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/Drawing.cpp
@@ -343,7 +343,7 @@ void Drawing::drawText(string text, int x, int y, rgba_t colour, int font, int a
 	float height = bbox.Upper().Y() - bbox.Lower().Y();
 	if (alignment != ALIGN_LEFT) {
 		if (alignment == ALIGN_CENTER)
-			xpos -= width*0.5;
+			xpos -= int(width*0.5);
 		else
 			xpos -= width;
 	}
